% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_greta_arrays.R
\name{create_greta_arrays}
\alias{create_greta_arrays}
\title{Create Greta Arrays}
\usage{
create_greta_arrays(x, y)
}
\arguments{
\item{x}{x data}

\item{y}{y data}
}
\value{
greta arrays
}
\description{
Create Greta Arrays
}
